requirements.txt
Django
python-dotenv
termcolor
validate-docbr
pytest
pytest-cov
psycopg2
psycopg2-binary
pylint-django
djangorestframework
markdown
django-filter
pyyaml
uritemplate
pygments
django-guardian
faker


# asgiref==3.2.10                                                                                                                                           
# astroid==2.4.2
# isort==4.3.21
# lazy-object-proxy==1.4.3
# mccabe==0.6.1
# pylint==2.5.3
# pytz==2020.1
# six==1.15.0
# sqlparse==0.3.1
# toml==0.10.1
# typed-ast==1.4.1
# wrapt==1.12.1
